[dsl] Provide lookups for items via "plural_noun named: 'name'"
Previously, was using the bare hash. Which went like this:

  carry_all.projects['awesome-cat']

Now, it goes like this:

  carry_all.project named: 'awesome-cat'

Also.. you used to get a list of projects known that you were getting
a hash and that you had to grab the values of that hash. To get a list
of projects you'd go like thie:

  carry_all.projects.values # known that projects returns a Hash

Now you go like this:

  carry_all.projects